requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
pytesseract
googletrans==4.0.0rc1
gTTS
langdetect
